+++ packages/btcstaking/src/lib.rs
@@ -1,4 +1,4 @@
-mod asig;
+mod adaptor_sig;
 pub mod scripts_utils;
 pub mod sig_verify;
 pub mod tx_verify;
