+++ packages/eots/src/lib.rs
@@ -1,6 +1,6 @@
 pub mod eots;
 pub mod error;
 
-pub use eots::{PubRand, PublicKey, SecRand, SecretKey, Signature};
+pub use eots::{tagged_hash, PubRand, PublicKey, SecRand, SecretKey, Signature, CHALLENGE_TAG};
 pub use error::Error;
 pub type Result<T> = std::result::Result<T, Error>;
